+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v6.8.2
 ------
 
+* Remove deprecated XML user menu. `<https://github.com/lsst-ts/LOVE-frontend/pull/729>`_
 * Remove deprecated openSplice version from CSCExpanded. `<https://github.com/lsst-ts/LOVE-frontend/pull/730>`_
 * Include bump test status result on BumpTests component. `<https://github.com/lsst-ts/LOVE-frontend/pull/728>`_
 * Remove custom weather forecast logic from generic plot components. `<https://github.com/lsst-ts/LOVE-frontend/pull/727>`_