+++ CHANGELOG.rst
@@ -2,9 +2,10 @@
 Version History
 ===============
 
-v6.8.2
+v6.9.0
 ------
 
+* Extend the CreateNightReport component to allow editing and sending not sent reports. `<https://github.com/lsst-ts/LOVE-frontend/pull/731>`_
 * Remove deprecated XML user menu. `<https://github.com/lsst-ts/LOVE-frontend/pull/729>`_
 * Remove deprecated openSplice version from CSCExpanded. `<https://github.com/lsst-ts/LOVE-frontend/pull/730>`_
 * Include bump test status result on BumpTests component. `<https://github.com/lsst-ts/LOVE-frontend/pull/728>`_